+++ Simple Game Library/CHANGES.md
@@ -3,6 +3,7 @@
 ## Version 1.3.0
 Date: *2016-04-??*
 
+- Added utility method `determineCommandKey` to get the command key for the current operating system.
 - Improved input system. The `keyDown` method provides in addition to the key code the corresponding character.
 - Added utility class to test if a character is printable.
 - Added more JavaDoc comments.